pyproject.toml
[project]
name = "social-media"
version = "0.1.0"
description = "Social Media Application using Django and Rest Framework"
authors = [
    {name = "asilbek",email = "englishisakovich@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "faker (>=40.15.0,<41.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
